fix corregir errores de compilacion

+++ backend/src/main/java/com/uade/tpejemplo/dto/request/SolicitudAumentoRequest.java
@@ -4,6 +4,8 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
+import java.math.BigDecimal;
+
 public class SolicitudAumentoRequest {
 
     @NotBlank(message = "El DNI del cliente es obligatorio")
@@ -11,7 +13,7 @@ public class SolicitudAumentoRequest {
 
     @NotNull(message = "El monto solicitado es obligatorio")
     @Positive(message = "El monto solicitado debe ser mayor a cero")
-    private Double montoSolicitado;
+    private BigDecimal montoSolicitado;
 
     public String getDniCliente() {
         return dniCliente;
@@ -21,11 +23,11 @@ public class SolicitudAumentoRequest {
         this.dniCliente = dniCliente;
     }
 
-    public Double getMontoSolicitado() {
+    public BigDecimal getMontoSolicitado() {
         return montoSolicitado;
     }
 
-    public void setMontoSolicitado(Double montoSolicitado) {
+    public void setMontoSolicitado(BigDecimal montoSolicitado) {
         this.montoSolicitado = montoSolicitado;
     }
 }
